requirements.txt
langchain
openai
langchain-openai
python-dotenv
crewai[tools]
streamlit
yfinance
plotly